+++ benchmarks/requests-1/run_sde.py
@@ -7,7 +7,8 @@
 # @Author:  LightningRS
 # @Email:   me@ldby.site
 # @Desc:    PyFuzzSDE Real Project Benchmark (requests-1) PyFuzzSDE Adapter
-# @Ref:     https://github.com/psf/requests/commit/ac4e05874a1a983ca126185a0e4d4e74915f792e
+# @BugAt:   https://github.com/psf/requests/commit/3bb13f8fbb9d4ed1a20bd33495cdc087eb062ca0
+# @Fixed:   https://github.com/psf/requests/commit/961790f95c7c06dc073d882acb28810a22d27b77
 # @Changelog:
 #    2021/05/18: Create benchmark
 
@@ -18,29 +19,40 @@ from pyfuzzsde.runner import SDERunner
 base_path = os.path.join(os.path.dirname(__file__), "requests")
 sys.path.append(base_path)
 
-from requests.utils import get_auth_from_url
-from requests.compat import quote
+from http.cookies import SimpleCookie
+from requests.cookies import morsel_to_cookie
 
 def requests_1_sde(one_input):
     try:
         ipt_data = str(one_input)
     except UnicodeDecodeError:
         sys.exit(0)
-    url = "http://{}:{}@test.com/test".format(
-        quote(ipt_data), quote(ipt_data)
-    )
-    usr, pwd = get_auth_from_url(url)
-    assert usr == ipt_data and pwd == ipt_data
+        
+    C = SimpleCookie(ipt_data)
+    for morsel in C.values():
+        cookie = morsel_to_cookie(morsel)
+        print(cookie)
+        
+    ipt_data2 = "{}={}; path=/; domain=.test.com".format(ipt_data, ipt_data)
+    C = SimpleCookie(ipt_data2)
+    for morsel in C.values():
+        cookie = morsel_to_cookie(morsel)
+        print(cookie)
 
 if __name__ == "__main__":
     runner = SDERunner(entry_func=requests_1_sde, trace_native_libs=False, 
         verbose=False, max_len=32,
         black_filenames=['PyFuzzSDE/pyfuzzsde'],
-        white_filenames=['urllib', 're.py']
+        white_filenames=['urllib', 'http']
     )
     # runner.set_entry_dir(base_path)
     runner.set_entry_dir('/')
-    runner.collect(times=500, corpus=['test'])
+    corpus = [
+        'sid=1; path=/; domain=.yahoo.com',
+        'PHPSESSID=aaa; path=/; domain=.yahoo.com',
+        'UID=123; path=/; domain=.yahoo.com',
+    ]
+    runner.collect(times=500, corpus=corpus)
 
     print(runner.used_data_pool)
     print(runner.data_pool)
@@ -54,8 +66,18 @@ if __name__ == "__main__":
     f_dict = open('./output/requests-1.dict', 'w')
     for i in range(len(final_pool)):
         src: str = final_pool[i]
-        trans = src.replace('\\', '\\\\').replace('"', '\\"')
+        trans = src.replace('\\', '\\\\').replace('"', '\\"').replace('\n', '')
         with open(os.path.join(all_input, 'input{}.txt'.format(i)), "w") as f:
             f.write(src)
-        f_dict.write('input{}="{}"\n'.format(i, trans))
+        if len(trans) > 0 and trans.isprintable():
+            f_dict.write('input{}="{}"\n'.format(i, trans))
     f_dict.close()
+
+    one_input = './output/requests-1-input1'
+    os.makedirs(one_input, exist_ok=True)
+    for i in range(len(corpus)):
+        src = corpus[i]
+        with open(os.path.join(one_input, 'input{}.txt'.format(i)), 'w') as f:
+            f.write(src)
+
+    runner.cov.report()