+++ benchmarks/requests-3/run_sde.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+export PYTHONPATH=$(pwd)/requests:$(dirname $(dirname $(pwd)))
+python run_sde.py
+exit 0